beta 0.2 cliente

+++ resources/views/clientes/script.blade.php
@@ -1,7 +1,7 @@
 <!-- Delay table load until everything else is loaded -->
 <script>
     $(window).load(function(){
-        $('#consultaTable').removeAttr('style');
+        $('#clienteTable').removeAttr('style');
     })
 </script>
 
@@ -39,14 +39,15 @@
     // add a new cliente
     $(document).on('click', '.add-modal', function() {
         $('.modal-title').text('Añadir Cliente');
-        $('#addModal').modal('show');
+        $('#add-modal').modal('show');
     });
     $('.modal-footer').on('click', '.add', function() {
         $.ajax({
             type: 'POST',
-            url: "clientes/",
+            url: "{{ URL::route('clientes.store') }}",
             data: {
                 '_token': $('input[name=_token]').val(),
+                'fecha': $('#fecha_add').val(),
                 'full_name': $('#full_name_add').val(),
                 'peso_inicial': $('#peso_inicial_add').val(),
                 'peso_saludable': $('#peso_saludable_add').val(),
@@ -54,7 +55,7 @@
                 'f_nacimiento': $('#f_nacimiento_add').val(),
                 'telefono': $('#telefono_add').val(),
                 'email': $('#email_add').val(),
-                'anotaciones': $('#anotaciones_add').val()
+                'anotaciones': $('#anotaciones_add').val(),
             },
             success: function(data) {
                 $('.errorFull_name').addClass('hidden');
@@ -68,7 +69,7 @@
 
                 if ((data.errors)) {
                     setTimeout(function () {
-                        $('#addModal').modal('show');
+                        $('#add-modal').modal('show');
                         toastr.error('Validation error!', 'Error Alert', {timeOut: 5000});
                     }, 500);
 
@@ -104,54 +105,61 @@
                         $('.errorAnotaciones').removeClass('hidden');
                         $('.errorAnotaciones').text(data.errors.anotaciones);
                     }
-                } else {   
-                    toastr.success('Cliente Añadido con Exito!', 'Success Alert', {timeOut: 5000});                             
-                    location.reload();                                
+                } else {  
+                    location.reload();  
+                    toastr.success('Cliente Añadido con Exito!', 'Success Alert', {timeOut: 5000});                                                            
                 }
+            }, 
+            error :function(error){
+                location.reload();
+                toastr.error('Agregation error! ('+eval(error)+')', 'Error Alert', {timeOut: 5000});
             },
-            error:function(){ 
-                alert("ERROR !!! Valores duplicados en el sistema, el teléfono y el email son únicos.");
-            
-            }
         });
+         
     });
 
-    // Show a post
+    // Show a cliente
     $(document).on('click', '.show-modal', function() {
         $('.modal-title').text('Detalles del Cliente Nº '+$(this).data('id'));
         $('#full_name_show').val($(this).data('full_name'));
-        $('#peso_show').val($(this).data('peso'));
+        $('#peso_inicial_show').val($(this).data('peso_inicial'));
+        $('#peso_saludable_show').val($(this).data('peso_saludable'));
         $('#f_nacimiento_show').val($(this).data('f_nacimiento'));
+        $('#altura_show').val($(this).data('altura'));
         $('#telefono_show').val($(this).data('telefono'));
         $('#email_show').val($(this).data('email'));
         $('#anotaciones_show').val($(this).data('anotaciones'));
-        $('#showModal').modal('show');
+        $('#show-modal').modal('show');
     });
 
-
-    // Edit a post
+    // Edit a cliente
     $(document).on('click', '.edit-modal', function() {
         id = $(this).data('id');
-        $('.modal-title').text('Editar Información del Cliente Nº '+id);
+        nombre = $(this).data('full_name');
+        $('.modal-title').text('Editar Información del Cliente (Nº '+id+') '+nombre);
         $('#id_edit').val(id);
-        $('#full_name_edit').val($(this).data('full_name'));
-        $('#peso_edit').val($(this).data('peso'));
+        $('#full_name_edit').val(nombre);
+        $('#peso_inicial_edit').val($(this).data('peso_inicial'));
+        $('#peso_saludable_edit').val($(this).data('peso_saludable'));
         $('#f_nacimiento_edit').val($(this).data('f_nacimiento'));
+        $('#altura_edit').val($(this).data('altura'));
         $('#telefono_edit').val($(this).data('telefono'));
         $('#email_edit').val($(this).data('email'));
         $('#anotaciones_edit').val($(this).data('anotaciones'));
-        $('#editModal').modal('show');
+        $('#edit-modal').modal('show');
     });
     id = $("#id_edit").val();
     $('.modal-footer').on('click', '.edit', function() {
         $.ajax({
             type: 'PUT',
-            url: '/clientes/' + id,
+            url: 'clientes/' + id,
             data: {
                 '_token': $('input[name=_token]').val(),
                 'id': $("#id_edit").val(),
                 'full_name': $('#full_name_edit').val(),
-                'peso': $('#peso_edit').val(),
+                'peso_inicial': $('#peso_inicial_edit').val(),
+                'peso_saludable': $('#peso_saludable_edit').val(),
+                'altura': $('#altura_edit').val(),
                 'f_nacimiento': $('#f_nacimiento_edit').val(),
                 'telefono': $('#telefono_edit').val(),
                 'email': $('#email_edit').val(),
@@ -159,7 +167,9 @@
             },
             success: function(data) {
                 $('.errorFull_name').addClass('hidden');
-                $('.errorPeso').addClass('hidden');
+                $('.errorPeso_inicial').addClass('hidden');
+                $('.errorPeso_saludable').addClass('hidden');
+                $('.errorAltura').addClass('hidden');
                 $('.errorF_nacimiento').addClass('hidden');
                 $('.errorTelefono').addClass('hidden');
                 $('.errorEmail').addClass('hidden');
@@ -167,7 +177,7 @@
 
                 if ((data.errors)) {
                     setTimeout(function () {
-                        $('#editModal').modal('show');
+                        $('#edit-modal').modal('show');
                         toastr.error('Validation error!', 'Error Alert', {timeOut: 5000});
                     }, 500);
 
@@ -175,9 +185,17 @@
                         $('.errorFull_name').removeClass('hidden');
                         $('.errorFull_name').text(data.errors.full_name);
                     }
-                    if (data.errors.peso) {
-                        $('.errorPeso').removeClass('hidden');
-                        $('.errorPeso').text(data.errors.peso);
+                    if (data.errors.peso_inicial) {
+                        $('.errorPeso_inicial').removeClass('hidden');
+                        $('.errorPeso_inicial').text(data.errors.peso_inicial);
+                    }
+                    if (data.errors.peso_saludable) {
+                        $('.errorPeso_saludable').removeClass('hidden');
+                        $('.errorPeso_saludable').text(data.errors.peso_saludable);
+                    }
+                    if (data.errors.altura) {
+                        $('.errorAltura').removeClass('hidden');
+                        $('.errorAltura').text(data.errors.altura);
                     }
                     if (data.errors.f_nacimiento) {
                         $('.errorF_nacimiento').removeClass('hidden');
@@ -197,11 +215,37 @@
                     }
                 } else {
                     location.reload();
-                    toastr.success('Successfully updated Post!', 'Success Alert', {timeOut: 5000});
+                    toastr.success('Successfully updated Cliente!', 'Success Alert', {timeOut: 5000});
                 }
-            }
+            },
+            error :function(error){
+                location.reload();
+                toastr.error('Edition error! ('+eval(error)+')', 'Error Alert', {timeOut: 5000});
+            },
         });
     });
+    /*
+    error: function (jqXHR, exception) {
+        var msg = '';
+        if (jqXHR.status === 0) {
+            msg = 'Not connect.\n Verify Network.';
+        } else if (jqXHR.status == 404) {
+            msg = 'Requested page not found. [404]';
+        } else if (jqXHR.status == 500) {
+            msg = 'Internal Server Error [500].';
+        } else if (exception === 'parsererror') {
+            msg = 'Requested JSON parse failed.';
+        } else if (exception === 'timeout') {
+            msg = 'Time out error.';
+        } else if (exception === 'abort') {
+            msg = 'Ajax request aborted.';
+        } else {
+            msg = 'Uncaught Error.\n' + jqXHR.responseText;
+        }
+        $('#post').html(msg);
+    },
+    */
+
 
     // delete a post
     $(document).on('click', '.delete-modal', function() {
@@ -209,21 +253,22 @@
         $('.modal-title').text('Borrar Ficha de Cliente Nº '+id);     
         $('#id_delete').val(id);
         $('#full_name_delete').val($(this).data('full_name'));
-        $('#deleteModal').modal('show');
+        $('#delete-modal').modal('show');
     });
     $('.modal-footer').on('click', '.delete', function() {
-        $.ajax({
-            type: 'DELETE',
-            url: 'clientes/'+ id,
-            data: {
-                '_token': $('input[name=_token]').val(),
-                'id': $("#id_delete").val(),
-            },
-            success: function(data) { 
-                location.reload();                              
-                toastr.success('Successfully deleted Post!', 'Success Alert', {timeOut: 5000});
-                //$('.item' + data['id']).remove();
-            }
-        });
+        if(confirm('¿Estas seguro de eliminar al cliente: '+$("#full_name_edit").val()+'?'))
+            $.ajax({
+                type: 'DELETE',
+                url: 'clientes/'+ id,
+                data: {
+                    '_token': $('input[name=_token]').val(),
+                    'id': $("#id_edit").val(),
+                },
+                success: function(data) { 
+                    location.reload();                              
+                    toastr.success('Successfully deleted Post!', 'Success Alert', {timeOut: 5000});
+                    //$('.item' + data['id']).remove();
+                }
+            });
     });
 </script>
